requirements.txt
pimux
uuid
requests
mysql-connector-python
streamlit
pandas
folium
plotly
numpy
matplotlib